+++ stplot-gui/modules/ReadWriteDevice/stlinkdevice.h
@@ -20,7 +20,7 @@ public:
     int initDevise(QVector<struct ReadAddres> readSeuqence);
     void stopDev();
     int execReadDevice();
-    int writeDataDevice(uint32_t data, varloc_location_t location);
+    int writeDataDevice(uint64_t data, varloc_location_t location);
     QWidget *getReadDevConfigWidget();
 
 private:
